requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
python-multipart==0.0.20
pydantic==2.11.7
python-dotenv==1.1.0
asyncpg==0.30.0
aiofiles==24.1.0
openai-whisper==20240930
torch==2.7.1
torchaudio==2.7.1
numpy==2.2.6
